+++ lib/discordrb/data/user.rb
@@ -143,8 +143,8 @@ module Discordrb
     # Add a blocking await for a message from this user. Specifically, this adds a global await for a MessageEvent with this
     # user's ID as a :from attribute.
     # @see Bot#add_await!
-    def await!(attributes = {})
-      @bot.add_await!(Discordrb::Events::MessageEvent, { from: @id }.merge(attributes))
+    def await!(attributes = {}, &block)
+      @bot.add_await!(Discordrb::Events::MessageEvent, { from: @id }.merge(attributes), &block)
     end
 
     # Gets the member this user is on a server